requirements.txt
pandas
pyyaml
requests
streamlit
yfinance>=0.2.30
numpy
